+++ src/parser/parser.rs
@@ -143,7 +143,7 @@ impl<'a> Parser<'a> {
     let tk = self.at().token_type.clone();
     match tk {
       lexer::TokenType::Ident => ast::Expr::Ident { symbol: self.consume().value },
-      lexer::TokenType::IntLit => ast::Expr::IntLit { value: self.consume().value.parse::<i32>().unwrap() },
+      lexer::TokenType::IntLit => ast::Expr::IntLit { value: self.consume().value.parse::<f32>().unwrap() },
       lexer::TokenType::OpenParen => {
         self.consume(); // consume the open paren
         let expr = self.parse_expr();
